backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
stripe==7.4.0
redis==5.0.1
huggingface-hub==0.19.4
transformers==4.35.2
torch==2.1.1
